+++ app/src/main/AndroidManifest.xml
@@ -9,15 +9,15 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".ui.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MainActivityConcursos" />
-        <activity android:name=".MainActivityEventos" />
+        <activity android:name=".ui.MainActivityConcursos" />
+        <activity android:name=".ui.MainActivityEventos" />
     </application>
 
 </manifest>
