+++ app/src/main/java/com/example/macbook/informatec/ui/MainActivity.java
@@ -1,24 +1,12 @@
-package com.example.macbook.informatec;
+package com.example.macbook.informatec.ui;
 
 import android.content.Intent;
-import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.widget.Button;
-import android.widget.Toast;
 
-import com.example.macbook.informatec.Control.MyAdapter;
-import com.example.macbook.informatec.Models.Events;
-import com.google.firebase.database.DataSnapshot;
-import com.google.firebase.database.DatabaseError;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.database.FirebaseDatabase;
-import com.google.firebase.database.ValueEventListener;
-
-import java.util.ArrayList;
+import com.example.macbook.informatec.R;
 
 public class MainActivity extends AppCompatActivity {
