+++ app/src/main/java/com/example/macbook/informatec/ui/MainActivityEventos.java
@@ -1,4 +1,4 @@
-package com.example.macbook.informatec;
+package com.example.macbook.informatec.ui;
 
 import android.support.annotation.NonNull;
 import android.support.v7.app.AppCompatActivity;
@@ -7,8 +7,9 @@ import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.widget.Toast;
 
-import com.example.macbook.informatec.Control.MyAdapter;
-import com.example.macbook.informatec.Models.Events;
+import com.example.macbook.informatec.data.network.MyAdapter;
+import com.example.macbook.informatec.data.models.Events;
+import com.example.macbook.informatec.R;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.DatabaseReference;
@@ -17,7 +18,7 @@ import com.google.firebase.database.ValueEventListener;
 
 import java.util.ArrayList;
 
-public class MainActivityConcursos extends AppCompatActivity {
+public class MainActivityEventos extends AppCompatActivity {
 
     RecyclerView recyclerView;
     DatabaseReference reference;
@@ -27,13 +28,13 @@ public class MainActivityConcursos extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main_concursos);
+        setContentView(R.layout.activity_main_eventos);
 
-        recyclerView =findViewById(R.id.myRecyclerView3);
+        recyclerView =findViewById(R.id.myRecyclerView2);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
         list = new ArrayList<Events>();
 
-        reference = FirebaseDatabase.getInstance().getReference().child("Robotica");
+        reference = FirebaseDatabase.getInstance().getReference().child("Eventos");
 
         reference.addValueEventListener(new ValueEventListener() {
             @Override
@@ -47,13 +48,13 @@ public class MainActivityConcursos extends AppCompatActivity {
 
                 }
 
-                adapter = new MyAdapter(MainActivityConcursos.this,list);
+                adapter = new MyAdapter(MainActivityEventos.this,list);
                 recyclerView.setAdapter(adapter);
             }
 
             @Override
             public void onCancelled(@NonNull DatabaseError databaseError) {
-                Toast.makeText(MainActivityConcursos.this,"Salio algo mal",Toast.LENGTH_LONG).show();
+                Toast.makeText(MainActivityEventos.this,"Salio algo mal",Toast.LENGTH_LONG).show();
             }
         });
 
@@ -61,3 +62,5 @@ public class MainActivityConcursos extends AppCompatActivity {
 
     }
 }
+
+
