+++ app/src/main/res/layout/activity_main_concursos.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivityConcursos">
+    tools:context=".ui.MainActivityConcursos">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/myRecyclerView3"
