+++ app/src/main/res/layout/activity_main_eventos.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivityEventos">
+    tools:context=".ui.MainActivityEventos">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/myRecyclerView2"
